Introduced Rubocop & first batch of fixes

+++ lib/sle2docker/prebuilt_image.rb
@@ -1,18 +1,19 @@
 module Sle2Docker
-
+  # This class takes care of handling the pre-build images for
+  # SUSE Linux Enterprise
   class PrebuiltImage
+    IMAGES_DIR = '/usr/share/suse-docker-images'
 
-    IMAGES_DIR = "/usr/share/suse-docker-images"
-
+    # Enum of the supported base systems
     class SUPPORTED_BASE_SYSTEMS
-      SLE11SP3 = "sle11sp3"
-      SLE12    = "sle12"
+      SLE11SP3 = 'sle11sp3'
+      SLE12    = 'sle12'
     end
 
     def self.list
-      if File.exists?(PrebuiltImage::IMAGES_DIR)
-        Dir[File.join(IMAGES_DIR, "*.tar.xz")].map do |image|
-          File.basename(image, ".tar.xz")
+      if File.exist?(PrebuiltImage::IMAGES_DIR)
+        Dir[File.join(IMAGES_DIR, '*.tar.xz')].map do |image|
+          File.basename(image, '.tar.xz')
         end
       else
         []
@@ -23,28 +24,27 @@ module Sle2Docker
       @image_name = image_name
       @options    = options
 
-      @base_system = if @image_name =~ /\Asle.12/
-        SUPPORTED_BASE_SYSTEMS::SLE12
+      if @image_name =~ /\Asle.12/
+        @base_system = SUPPORTED_BASE_SYSTEMS::SLE12
       elsif @image_name =~ /\Asle.11sp3/
-        SUPPORTED_BASE_SYSTEMS::SLE11SP3
+        @base_system = SUPPORTED_BASE_SYSTEMS::SLE11SP3
       end
     end
 
     def activate
-      if !File.exists?(File.join(IMAGES_DIR, "#{@image_name}.tar.xz"))
-        raise PrebuiltImageNotFoundError.new("Cannot find pre-built image #{@image_name}")
+      unless File.exist?(File.join(IMAGES_DIR, "#{@image_name}.tar.xz"))
+        fail PrebuiltImageNotFoundError,
+             "Cannot find pre-built image #{@image_name}"
       end
 
       verify_image
 
       tmp_dir = prepare_docker_build_root
-      puts "Activating image"
+      puts 'Activating image'
       image = Docker::Image.build_from_dir(tmp_dir)
-      image.tag(docker_tag())
+      image.tag(docker_tag)
     ensure
-      if tmp_dir && File.exists?(tmp_dir)
-        FileUtils.rm_rf(tmp_dir)
-      end
+      FileUtils.rm_rf(tmp_dir) if tmp_dir && File.exist?(tmp_dir)
     end
 
     def prepare_docker_build_root
@@ -58,6 +58,7 @@ module Sle2Docker
       tmp_dir
     end
 
+    # rubocop:disable Lint/UselessAssignment
     def create_dockerfile(tmp_dir)
       repositories  = {}
       credentials   = {}
@@ -66,38 +67,40 @@ module Sle2Docker
       case @base_system
       when SUPPORTED_BASE_SYSTEMS::SLE12
         host = @options[:smt_host]
-        template_name = "sle12-dockerfile.erb"
+        template_name = 'sle12-dockerfile.erb'
       when SUPPORTED_BASE_SYSTEMS::SLE11SP3
-        host = @options[:smt_host] || "nu.novell.com"
-        if @options[:smt_host].nil? 
+        host = @options[:smt_host] || 'nu.novell.com'
+        if @options[:smt_host].nil?
           cred_helper = CredentialsHelper.new(@options, true)
           credentials[:username] = cred_helper.username
           credentials[:password] = cred_helper.password
-          credentials[:filename] = "NCCcredentials"
+          credentials[:filename] = 'NCCcredentials'
         end
 
-        template_name = "sle11sp3-dockerfile.erb"
+        template_name = 'sle11sp3-dockerfile.erb'
       else
-        raise TemplateNotFoundError.new("Cannot find right template for #{@image_name}")
+        fail TemplateNotFoundError,
+             "Cannot find right template for #{@image_name}"
       end
 
-      prebuilt_image = @image_name + ".tar.xz"
+      prebuilt_image = @image_name + '.tar.xz'
 
       template = ERB.new(
         File.read(
           File.join(
-            File.expand_path("../../templates/docker_build", __FILE__),
+            File.expand_path('../../templates/docker_build', __FILE__),
             template_name
           )
         ),
         nil,
-        "<>"
+        '<>'
       ).result(binding)
 
-      File.open(File.join(tmp_dir, "Dockerfile"), "w") do |file|
+      File.open(File.join(tmp_dir, 'Dockerfile'), 'w') do |file|
         file.write(template)
       end
     end
+    # rubocop:enable Lint/UselessAssignment
 
     def copy_prebuilt_image(tmp_dir)
       prebuilt_image = File.join(IMAGES_DIR, "#{@image_name}.tar.xz")
@@ -109,40 +112,41 @@ module Sle2Docker
       repos_to_copy = []
 
       # copy services.d files
-      source = "/etc/zypp/services.d"
+      source = '/etc/zypp/services.d'
       services = Dir["#{source}/*.service"]
       if services.empty?
-        raise ConfigNotFoundError.new("No service file found under /etc/zypp/services.d")
+        fail ConfigNotFoundError,
+             'No service file found under /etc/zypp/services.d'
       end
-      FileUtils.mkdir_p(File.join(tmp_dir, "zypp", "services.d"))
+      FileUtils.mkdir_p(File.join(tmp_dir, 'zypp', 'services.d'))
       services.each do |service|
         FileUtils.cp(
           service,
-          File.join(tmp_dir, "zypp", "services.d", File.basename(service))
+          File.join(tmp_dir, 'zypp', 'services.d', File.basename(service))
         )
         repos_to_copy += File.readlines(service)
-          .grep(/repo_\d+=/)
-          .map{ |match| match.split("=", 2)[1].chomp + ".repo" }
+                             .grep(/repo_\d+=/)
+                             .map { |match| match.split('=', 2)[1].chomp + '.repo' }
       end
 
       # copy credentials.d files
-      source = "/etc/zypp/credentials.d"
+      source = '/etc/zypp/credentials.d'
       if !File.exist?(source)
-        raise ConfigNotFoundError.new("#{source} does not exist")
+        fail ConfigNotFoundError, "#{source} does not exist"
       else
-        FileUtils.cp_r(source, File.join(tmp_dir, "zypp"))
+        FileUtils.cp_r(source, File.join(tmp_dir, 'zypp'))
       end
 
       # copy repositories
-      destination = File.join(tmp_dir, "zypp", "repos.d")
+      destination = File.join(tmp_dir, 'zypp', 'repos.d')
       FileUtils.mkdir_p(destination)
 
       repos_to_copy.each do |repo|
-        source = File.join("/etc/zypp/repos.d", repo)
-        if File.exists?(source)
+        source = File.join('/etc/zypp/repos.d', repo)
+        if File.exist?(source)
           FileUtils.cp(source, File.join(destination, repo))
         else
-          raise ConfigNotFoundError.new("Cannot find repository file #{source}")
+          fail ConfigNotFoundError, "Cannot find repository file #{source}"
         end
       end
     end
@@ -151,30 +155,30 @@ module Sle2Docker
       # example of image name: sles12-docker.x86_64-1.0.0-Build7.2
       match = /\A(?<name>.*)-docker\..*-(?<version>\d+\.\d+\.\d+)/.match(@image_name)
       if match.nil?
-        raise DockerTagError.new("Cannot calculate the Docker tag for #{@image_name}")
+        fail DockerTagError,
+             "Cannot calculate the Docker tag for #{@image_name}"
       end
 
       {
-        "repo" => "suse/#{match["name"]}",
-        "tag" => match["version"]
+        'repo' => "suse/#{match['name']}",
+        'tag' => match['version']
       }
     end
 
     def verify_image
       file = File.join(IMAGES_DIR, "#{@image_name}.tar.xz")
       package_name = `rpm -qf #{file}`
-      if $?.exitstatus != 0
-        raise PrebuiltImageVerificationError.new("Cannot find rpm package providing #{file}: #{package_name}")
+      if $CHILD_STATUS.exitstatus != 0
+        fail PrebuiltImageVerificationError,
+             "Cannot find rpm package providing #{file}: #{package_name}"
       end
 
-      puts "Verifying integrity of the pre-built image"
+      puts 'Verifying integrity of the pre-built image'
       verification = `rpm --verify #{package_name}`
-      if $?.exitstatus != 0
-        raise PrebuiltImageVerificationError.new("Verification of #{package_name} failed: #{verification}")
+      if $CHILD_STATUS.exitstatus != 0
+        fail PrebuiltImageVerificationError,
+             "Verification of #{package_name} failed: #{verification}"
       end
     end
-
   end
-
 end
-
